# Improve char.ToUpperInvariant / ToLowerInvariant perf