@GrabYourPitchforks 2020-04-20
I also tried testing a branchless implementation but the results were about the same for the different inputs I tried. So no need to complicate the logic.