@gfoidl 2020-04-20
Should the above benchmarks be added to [dotnet/performance](https://github.com/dotnet/performance/blob/8aed638c9ee65c034fe0cca4ea2bdc3a68d2a6b5/src/benchmarks/micro/libraries/System.Runtime/Perf.Char.cs#L13)?
For `string` there are benchmarks for `ToUpperInvariant`, but not for `char`.